old-tests/submission/W3C/domhtml/HTMLObjectElement17.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/dimension-attributes.html#dom-dim-width" rel="help">
<link rel="author" title="NIST">
</head>
<body>

<object data="/support/images/yellow-rectangle.png" height="60" type="image/png" width="550"></object>



<script>
var domtest = async_test("The width attribute overrides the original width value.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vwidth;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vwidth = testNode.width;

      assert_equals(vwidth,"550");

      domtest.done();
}));

</script>
<p>The width attribute overrides the original width value.                

    Retrieve the width attribute of the first OBJECT element and examine
    its value.</p>
<div id="log"></div>
</body></html>
